fix: include header to compare floats with Catch2

+++ include/average.h
@@ -1,6 +1,6 @@
 #ifndef _AVERAGE_H_
 #define _AVERAGE_H_
 
-int average(int list[], int n);
+double average(int list[], int n);
 
 #endif // _AVERAGE_H_
